centralize swap execution state
