+++ packages/widget/src/pages/SwapExecutionPage/useSwapExecutionState.ts
@@ -2,7 +2,10 @@
 import { useMemo } from "react";
 import { ChainAddress } from "@/state/swapExecutionPage";
 import { SimpleStatus } from "@/utils/clientType";
-import { SwapExecutionState } from "./SwapExecutionPage";
+import {
+  SwapExecutionState,
+  computeSwapExecutionState,
+} from "@/utils/swapExecutionState";
 import { RouteResponse } from "@skip-go/client";
 
 type UseSwapExecutionStateParams = {
@@ -22,56 +25,23 @@ export function useSwapExecutionState({
   signaturesRemaining,
   isLoading,
 }: UseSwapExecutionStateParams): SwapExecutionState {
-  return useMemo(() => {
-    if (isLoading) return SwapExecutionState.pendingGettingAddresses;
-    if (!chainAddresses) return SwapExecutionState.destinationAddressUnset;
-    const requiredChainAddresses = route?.requiredChainAddresses;
-    if (!requiredChainAddresses) return SwapExecutionState.destinationAddressUnset;
-
-    const allAddressesSet = requiredChainAddresses.every(
-      (_chainId, index) => chainAddresses[index]?.address,
-    );
-
-    const lastChainAddress = chainAddresses[requiredChainAddresses.length - 1]?.address;
-
-    if (overallStatus === "completed") {
-      return SwapExecutionState.confirmed;
-    }
-
-    if (overallStatus === "pending") {
-      if (signaturesRemaining > 0) {
-        return SwapExecutionState.signaturesRemaining;
-      }
-      return SwapExecutionState.pending;
-    }
-
-    if (overallStatus === "approving") {
-      return SwapExecutionState.approving;
-    }
-
-    if (isValidatingGasBalance && isValidatingGasBalance.status !== "completed") {
-      return SwapExecutionState.validatingGasBalance;
-    }
-
-    if (overallStatus === "signing") {
-      return SwapExecutionState.waitingForSigning;
-    }
-
-    if (!lastChainAddress) {
-      return SwapExecutionState.destinationAddressUnset;
-    }
-
-    if (!allAddressesSet) {
-      return SwapExecutionState.recoveryAddressUnset;
-    }
-
-    return SwapExecutionState.ready;
-  }, [
-    isLoading,
-    chainAddresses,
-    route?.requiredChainAddresses,
-    overallStatus,
-    isValidatingGasBalance,
-    signaturesRemaining,
-  ]);
+  return useMemo(
+    () =>
+      computeSwapExecutionState({
+        chainAddresses,
+        route,
+        overallStatus,
+        isValidatingGasBalance,
+        signaturesRemaining,
+        isLoading,
+      }),
+    [
+      isLoading,
+      chainAddresses,
+      route?.requiredChainAddresses,
+      overallStatus,
+      isValidatingGasBalance,
+      signaturesRemaining,
+    ],
+  );
 }
